refactor(cli): use shared facades for JSON helpers
