+++ src/cli/lib/shared/json-utils.js
@@ -1,0 +1,1 @@
+export * from "../../../shared/json-utils.js";
